requirements.txt
inquirer==3.2.4
pyfiglet==1.0.2
python-nmap==0.7.1
cryptography==42.0.5
